requirements.txt
torch
transformers
numpy
opencv-python
sounddevice
streamlit
gtts
pydub
SpeechRecognition
deepface
python-dotenv
groq
google-cloud-speech
langchain
langchain-community
langchain-huggingface
faiss-cpu
PyMuPDF
